requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
openai==1.3.0
python-dotenv==1.0.0
markdown==3.5.1
Pillow
google-generativeai==0.3.1
gunicorn
whitenoise